removing unnecessary casts...
git-svn-id: https://svn.java.net/svn/glazedlists~svn/trunk@2274 696c4f5c-992c-4387-df9e-f6c076136008

+++ test/ca/odell/glazedlists/calculation/SumTest.java
@@ -108,7 +108,7 @@ public final class SumTest extends TestCase {
 
         // test remove
         source.remove(1);
-        assertEquals(1, (int) sum.getValue().intValue());
+        assertEquals(1, sum.getValue().intValue());
         assertEquals(1, counter.getCountAndReset());
     }
